fix: Fix tests
